+++ api/format/single_elimination_test.go
@@ -7,6 +7,7 @@ package format
 import (
 	"testing"
 
+	"pickems-bot/api/external"
 	"pickems-bot/api/shared"
 
 	"github.com/stretchr/testify/assert"
@@ -151,3 +152,220 @@ func TestSingleElim_RequiredPredictions(t *testing.T) {
 		assert.Equal(t, c.want, singleElimFormat{}.RequiredPredictions(c.teams), "teamCount=%d", c.teams)
 	}
 }
+
+// region getEliminationResults
+
+// TestGetEliminationResults_GrandFinal tests single match (Grand Final)
+func TestGetEliminationResults_GrandFinal(t *testing.T) {
+	matchNodes := []external.MatchNode{
+		{ID: "bracket_R01-M001", Team1: "Winner", Team2: "RunnerUp", Winner: "Winner"},
+	}
+	results, err := getEliminationResults(matchNodes)
+	assert.NoError(t, err)
+	assert.Equal(t, "Grand Final", results["Winner"].Round)
+	assert.Equal(t, "advanced", results["Winner"].Status)
+	assert.Equal(t, "Grand Final", results["RunnerUp"].Round)
+	assert.Equal(t, "eliminated", results["RunnerUp"].Status)
+}
+
+// TestGetEliminationResults_MultipleRounds tests multiple rounds
+func TestGetEliminationResults_MultipleRounds(t *testing.T) {
+	matchNodes := []external.MatchNode{
+		{ID: "bracket_R02-M001", Team1: "TeamA", Team2: "TeamB", Winner: "TeamA"},
+		{ID: "bracket_R02-M002", Team1: "TeamC", Team2: "TeamD", Winner: "TeamC"},
+		{ID: "bracket_R01-M001", Team1: "TeamA", Team2: "TeamC", Winner: "TeamA"},
+	}
+	results, err := getEliminationResults(matchNodes)
+	assert.NoError(t, err)
+	assert.Len(t, results, 4)
+	assert.Equal(t, "advanced", results["TeamA"].Status)
+	assert.Equal(t, "eliminated", results["TeamC"].Status)
+	assert.Equal(t, "eliminated", results["TeamB"].Status)
+	assert.Equal(t, "eliminated", results["TeamD"].Status)
+}
+
+// TestGetEliminationResults_PendingMatches tests with TBD winners
+func TestGetEliminationResults_PendingMatches(t *testing.T) {
+	matchNodes := []external.MatchNode{
+		{ID: "bracket_R01-M001", Team1: "TeamA", Team2: "TeamB", Winner: "TBD"},
+	}
+	results, err := getEliminationResults(matchNodes)
+	assert.NoError(t, err)
+	assert.Equal(t, "Grand Final", results["TeamA"].Round)
+	assert.Equal(t, "pending", results["TeamA"].Status)
+	assert.Equal(t, "Grand Final", results["TeamB"].Round)
+	assert.Equal(t, "pending", results["TeamB"].Status)
+}
+
+// TestGetEliminationResults_EmptyWinner tests with empty winner string
+func TestGetEliminationResults_EmptyWinner(t *testing.T) {
+	matchNodes := []external.MatchNode{
+		{ID: "bracket_R01-M001", Team1: "TeamA", Team2: "TeamB", Winner: ""},
+	}
+	results, err := getEliminationResults(matchNodes)
+	assert.NoError(t, err)
+	assert.Equal(t, "pending", results["TeamA"].Status)
+	assert.Equal(t, "pending", results["TeamB"].Status)
+}
+
+// TestGetEliminationResults_EmptyInput tests error with no matches
+func TestGetEliminationResults_EmptyInput(t *testing.T) {
+	_, err := getEliminationResults([]external.MatchNode{})
+	assert.Error(t, err)
+	assert.Contains(t, err.Error(), "at least one match required")
+}
+
+// TestGetEliminationResults_EmptyTeamNames tests handling of empty team names
+func TestGetEliminationResults_EmptyTeamNames(t *testing.T) {
+	matchNodes := []external.MatchNode{
+		{ID: "bracket_R01-M001", Team1: "", Team2: "TeamB", Winner: "TeamB"},
+	}
+	results, err := getEliminationResults(matchNodes)
+	assert.NoError(t, err)
+	assert.Len(t, results, 1)
+	assert.Equal(t, "Grand Final", results["TeamB"].Round)
+	assert.Equal(t, "advanced", results["TeamB"].Status)
+}
+
+// endregion
+
+// region getRoundIndex
+
+func TestGetRoundIndex_Found(t *testing.T) {
+	rounds := []string{"Grand Final", "Semi Final", "Quarter Final"}
+	assert.Equal(t, 2, getRoundIndex("Semi Final", rounds))
+}
+
+func TestGetRoundIndex_FirstRound(t *testing.T) {
+	rounds := []string{"Grand Final", "Semi Final", "Quarter Final"}
+	assert.Equal(t, 3, getRoundIndex("Grand Final", rounds))
+}
+
+func TestGetRoundIndex_LastRound(t *testing.T) {
+	rounds := []string{"Grand Final", "Semi Final", "Quarter Final"}
+	assert.Equal(t, 1, getRoundIndex("Quarter Final", rounds))
+}
+
+func TestGetRoundIndex_NotFound(t *testing.T) {
+	assert.Equal(t, -1, getRoundIndex("Quarter Final", []string{"Grand Final", "Semi Final"}))
+}
+
+func TestGetRoundIndex_EmptyRounds(t *testing.T) {
+	assert.Equal(t, -1, getRoundIndex("Grand Final", []string{}))
+}
+
+// endregion
+
+// region getRoundNames
+
+func TestGetRoundNames_SingleMatch(t *testing.T) {
+	rounds, err := getRoundNames(1)
+	assert.NoError(t, err)
+	assert.Equal(t, []string{"Grand Final"}, rounds)
+}
+
+func TestGetRoundNames_ThreeMatches(t *testing.T) {
+	rounds, err := getRoundNames(3)
+	assert.NoError(t, err)
+	assert.Equal(t, []string{"Grand Final", "Semi Final"}, rounds)
+}
+
+func TestGetRoundNames_SevenMatches(t *testing.T) {
+	rounds, err := getRoundNames(7)
+	assert.NoError(t, err)
+	assert.Equal(t, []string{"Grand Final", "Semi Final", "Quarter Final"}, rounds)
+}
+
+func TestGetRoundNames_FifteenMatches(t *testing.T) {
+	rounds, err := getRoundNames(15)
+	assert.NoError(t, err)
+	assert.Equal(t, []string{"Grand Final", "Semi Final", "Quarter Final", "Best of 16"}, rounds)
+}
+
+func TestGetRoundNames_ThirtyOneMatches(t *testing.T) {
+	rounds, err := getRoundNames(31)
+	assert.NoError(t, err)
+	assert.Equal(t, []string{"Grand Final", "Semi Final", "Quarter Final", "Best of 16", "Best of 32"}, rounds)
+}
+
+func TestGetRoundNames_TooManyMatches(t *testing.T) {
+	_, err := getRoundNames(63)
+	assert.Error(t, err)
+	assert.Contains(t, err.Error(), "unsupported depth")
+}
+
+func TestGetRoundNames_ZeroMatches(t *testing.T) {
+	rounds, err := getRoundNames(0)
+	assert.NoError(t, err)
+	assert.Empty(t, rounds)
+}
+
+// endregion
+
+// region extractRoundAndMatchIDs
+
+func TestExtractRoundAndMatchIDs_Valid(t *testing.T) {
+	round, match, err := extractRoundAndMatchIDs("RSTxQ88PoQ_R03-M001")
+	assert.NoError(t, err)
+	assert.Equal(t, 3, round)
+	assert.Equal(t, 1, match)
+}
+
+func TestExtractRoundAndMatchIDs_DifferentValues(t *testing.T) {
+	round, match, err := extractRoundAndMatchIDs("ABC123_R10-M025")
+	assert.NoError(t, err)
+	assert.Equal(t, 10, round)
+	assert.Equal(t, 25, match)
+}
+
+func TestExtractRoundAndMatchIDs_InvalidFormat(t *testing.T) {
+	_, _, err := extractRoundAndMatchIDs("invalid_format")
+	assert.Error(t, err)
+	assert.Contains(t, err.Error(), "invalid ID format")
+}
+
+func TestExtractRoundAndMatchIDs_MissingUnderscore(t *testing.T) {
+	_, _, err := extractRoundAndMatchIDs("R03-M001")
+	assert.Error(t, err)
+}
+
+func TestExtractRoundAndMatchIDs_MissingRound(t *testing.T) {
+	_, _, err := extractRoundAndMatchIDs("bracket_M001")
+	assert.Error(t, err)
+}
+
+// endregion
+
+// region ExtractMatchListIDs
+
+// TestSingleElimExtractMatchListIDs_Basic tests Bracket ID extraction
+func TestSingleElimExtractMatchListIDs_Basic(t *testing.T) {
+	wikitext := `
+== Format ==
+Single-elimination bracket
+
+{{Bracket|id=BRACKET1}}
+`
+	ids, kind, err := singleElimFormat{}.ExtractMatchListIDs(wikitext)
+	assert.NoError(t, err)
+	assert.Equal(t, SingleElim, kind)
+	assert.Len(t, ids, 1)
+	assert.Contains(t, ids, "BRACKET1")
+}
+
+// TestSingleElimExtractMatchListIDs_StripsHTMLComments verifies trailing
+// HTML comments are removed from extracted IDs.
+func TestSingleElimExtractMatchListIDs_StripsHTMLComments(t *testing.T) {
+	wikitext := `
+== Format ==
+Single-elimination bracket
+
+{{Bracket|id=CLEAN123<!-- some comment -->}}
+`
+	ids, _, err := singleElimFormat{}.ExtractMatchListIDs(wikitext)
+	assert.NoError(t, err)
+	assert.Len(t, ids, 1)
+	assert.Equal(t, "CLEAN123", ids[0])
+}
+
+// endregion
